added custom method for addData
