+++ examples/timeSeries-mousexy-keypoints/sketch.js
@@ -1,104 +1,180 @@
+// https://editor.p5js.org/gohai/sketches/_KdpDSQzH
+
+let model;
+
+let curr_shape = 'None, press a button below'
+
+let state = 'collection';
+let pressedOnce = true;
+
+
+let rec_duration = 2;
+let num_seq = 20;
+// assuming frameRate is 60, with record time of 2 seconds, there will be 120 datapoints total, which is huge! we use map to get 20 data points instead of 120
+
+let count = 0;
+
 
 let sequence = [];
-let datasets = [];
-
-let classifer;
-let tensorData;
-
-let pressedOnce = true;
-let w =400 
-let h = 400;
-let type = ["box","circle"]
-
-let circle = [[[0.48,0.165],[0.555,0.175],[0.615,0.1875],[0.715,0.2475],[0.78,0.35],[0.7825,0.545],[0.7225,0.6375],[0.615,0.7175],[0.5175,0.74],[0.395,0.715],[0.25,0.6425],[0.2,0.5375],[0.205,0.44],[0.2425,0.345],[0.3,0.245],[0.3425,0.2025],[0.43,0.1675],[0.495,0.165],[0.5025,0.165],[0.5025,0.165]],[[0.47,0.16],[0.4925,0.1575],[0.6,0.1725],[0.6875,0.235],[0.7925,0.35],[0.8275,0.455],[0.7825,0.585],[0.715,0.695],[0.5825,0.725],[0.3725,0.695],[0.2275,0.6075],[0.1875,0.4775],[0.205,0.3975],[0.2625,0.3],[0.3525,0.21],[0.39,0.18],[0.4275,0.17],[0.4625,0.16],[0.4725,0.1575],[0.475,0.1575]],[[0.4675,0.145],[0.5375,0.15],[0.64,0.205],[0.7375,0.2925],[0.7875,0.4275],[0.7625,0.53],[0.6725,0.65],[0.55,0.7025],[0.4025,0.6975],[0.2975,0.6325],[0.25,0.5375],[0.24,0.4225],[0.265,0.3575],[0.3125,0.2875],[0.3675,0.225],[0.4025,0.1875],[0.4225,0.1625],[0.45,0.1475],[0.4675,0.1475],[0.4775,0.1475]],[[0.5525,0.1475],[0.6825,0.195],[0.7825,0.2775],[0.83,0.39],[0.8225,0.5075],[0.7825,0.62],[0.705,0.6925],[0.4875,0.76],[0.315,0.7625],[0.23,0.7125],[0.2025,0.605],[0.21,0.4625],[0.2575,0.3125],[0.36,0.1875],[0.4625,0.1425],[0.51,0.14],[0.5375,0.14],[0.5525,0.14],[0.5625,0.14],[0.5625,0.14]],[[0.5725,0.1325],[0.6825,0.155],[0.7625,0.2275],[0.8275,0.3525],[0.8325,0.48],[0.81,0.5925],[0.7675,0.685],[0.6425,0.735],[0.4725,0.76],[0.3375,0.715],[0.2775,0.625],[0.27,0.5225],[0.27,0.4325],[0.3025,0.35],[0.3625,0.27],[0.42,0.1975],[0.4625,0.165],[0.5175,0.1375],[0.5575,0.1275],[0.57,0.1275]],[[0.5225,0.1425],[0.62,0.1375],[0.7725,0.19],[0.8325,0.2775],[0.87,0.4275],[0.8375,0.6075],[0.81,0.67],[0.755,0.7075],[0.5525,0.715],[0.4175,0.7075],[0.2775,0.5975],[0.2625,0.48],[0.275,0.405],[0.3225,0.31],[0.3625,0.2525],[0.3925,0.2175],[0.425,0.1875],[0.4675,0.1575],[0.49,0.1475],[0.5025,0.145]],[[0.4875,0.135],[0.6625,0.1525],[0.765,0.2175],[0.8325,0.345],[0.8525,0.495],[0.845,0.64],[0.805,0.7075],[0.715,0.7525],[0.575,0.7775],[0.45,0.7775],[0.3475,0.7375],[0.285,0.675],[0.2625,0.555],[0.2825,0.4125],[0.32,0.315],[0.3625,0.2275],[0.3925,0.1975],[0.425,0.175],[0.4425,0.1675],[0.455,0.1575]],[[0.4025,0.1675],[0.455,0.1525],[0.5625,0.1525],[0.665,0.185],[0.7625,0.2475],[0.8225,0.355],[0.8475,0.5075],[0.83,0.6475],[0.7425,0.725],[0.5775,0.7925],[0.435,0.79],[0.2775,0.7325],[0.2075,0.6275],[0.2025,0.53],[0.2025,0.4525],[0.2225,0.36],[0.2575,0.3],[0.3,0.2475],[0.33,0.225],[0.36,0.1975]],[[0.4925,0.1525],[0.5925,0.1425],[0.69,0.1825],[0.7825,0.2425],[0.8325,0.3125],[0.87,0.45],[0.8825,0.6275],[0.8575,0.76],[0.7825,0.82],[0.5825,0.865],[0.435,0.865],[0.3525,0.8275],[0.2825,0.7075],[0.2675,0.6075],[0.2775,0.4825],[0.2975,0.3925],[0.335,0.31],[0.3725,0.2375],[0.4025,0.2025],[0.43,0.1775]],[[0.4725,0.15],[0.5825,0.1475],[0.6875,0.2],[0.7775,0.2875],[0.82,0.4],[0.8225,0.5575],[0.7525,0.7075],[0.665,0.7725],[0.555,0.8025],[0.395,0.7675],[0.305,0.68],[0.255,0.56],[0.245,0.42],[0.2675,0.34],[0.3325,0.235],[0.365,0.2],[0.4025,0.1725],[0.445,0.15],[0.475,0.1425],[0.5125,0.1425]],[[0.485,0.1375],[0.6175,0.1375],[0.73,0.205],[0.795,0.2875],[0.8175,0.4275],[0.8125,0.5525],[0.7825,0.635],[0.675,0.7125],[0.52,0.7425],[0.33,0.6975],[0.2625,0.5975],[0.2525,0.4975],[0.2675,0.3675],[0.2975,0.2775],[0.3425,0.2175],[0.4125,0.17],[0.4525,0.155],[0.495,0.14],[0.5125,0.14],[0.515,0.14]],[[0.5125,0.1325],[0.6425,0.15],[0.7425,0.225],[0.81,0.43],[0.81,0.6025],[0.74,0.7875],[0.6075,0.8425],[0.4375,0.86],[0.3,0.8125],[0.225,0.71],[0.1875,0.605],[0.2125,0.4825],[0.2625,0.3775],[0.3575,0.2625],[0.4025,0.205],[0.4425,0.1675],[0.515,0.1425],[0.5525,0.135],[0.5625,0.135],[0.5625,0.135]],[[0.5375,0.1375],[0.6275,0.16],[0.7475,0.235],[0.8025,0.3775],[0.8225,0.5725],[0.77,0.7225],[0.6675,0.835],[0.5475,0.8675],[0.38,0.835],[0.2525,0.7],[0.2125,0.5675],[0.2225,0.4375],[0.2625,0.3275],[0.3225,0.2425],[0.395,0.1875],[0.475,0.1625],[0.5275,0.15],[0.5425,0.1475],[0.5525,0.145],[0.5575,0.145]],[[0.53,0.1375],[0.635,0.1525],[0.725,0.2175],[0.8225,0.39],[0.825,0.6025],[0.76,0.79],[0.6625,0.8375],[0.545,0.84],[0.4075,0.8225],[0.3425,0.775],[0.285,0.6675],[0.28,0.5375],[0.295,0.3975],[0.345,0.315],[0.3775,0.2675],[0.4175,0.2175],[0.4575,0.1675],[0.4875,0.1475],[0.5075,0.1375],[0.52,0.1375]],[[0.4775,0.165],[0.5475,0.1575],[0.63,0.1775],[0.6675,0.2025],[0.6975,0.235],[0.7375,0.3225],[0.7525,0.465],[0.7225,0.5775],[0.655,0.6525],[0.5525,0.67],[0.425,0.655],[0.3275,0.59],[0.28,0.4775],[0.275,0.35],[0.305,0.2825],[0.34,0.235],[0.3775,0.1975],[0.455,0.1575],[0.475,0.15],[0.495,0.1475]],[[0.4375,0.165],[0.5425,0.1475],[0.635,0.1725],[0.685,0.2175],[0.735,0.305],[0.7475,0.4],[0.7125,0.4825],[0.64,0.5375],[0.5325,0.555],[0.3825,0.535],[0.3,0.48],[0.2825,0.41],[0.2725,0.3325],[0.2775,0.2825],[0.3,0.245],[0.34,0.2075],[0.3575,0.195],[0.4,0.17],[0.4225,0.1575],[0.465,0.1475]],[[0.4625,0.145],[0.5575,0.15],[0.6575,0.1975],[0.7275,0.2625],[0.755,0.3525],[0.7425,0.4775],[0.6825,0.595],[0.595,0.635],[0.375,0.64],[0.275,0.6175],[0.205,0.54],[0.1825,0.4075],[0.1925,0.3025],[0.2325,0.255],[0.2825,0.22],[0.3525,0.1775],[0.3875,0.1575],[0.4325,0.1425],[0.4725,0.14],[0.5025,0.1475]],[[0.4625,0.1175],[0.5325,0.1125],[0.6425,0.1425],[0.715,0.21],[0.765,0.3175],[0.7775,0.46],[0.7325,0.5775],[0.6175,0.63],[0.4775,0.655],[0.3675,0.635],[0.2725,0.58],[0.2075,0.4875],[0.22,0.385],[0.275,0.2725],[0.3525,0.2],[0.3825,0.1675],[0.42,0.1375],[0.455,0.1175],[0.475,0.1125],[0.4925,0.1125]],[[0.5425,0.125],[0.7125,0.165],[0.7925,0.26],[0.8275,0.415],[0.8175,0.5225],[0.725,0.6125],[0.615,0.675],[0.465,0.7175],[0.3725,0.6825],[0.32,0.6225],[0.2975,0.515],[0.31,0.4],[0.3325,0.33],[0.37,0.2775],[0.4275,0.225],[0.465,0.1925],[0.535,0.1375],[0.5625,0.1225],[0.5725,0.12],[0.5725,0.12]],[[0.505,0.1225],[0.5725,0.13],[0.6825,0.1975],[0.74,0.2925],[0.7375,0.485],[0.67,0.5875],[0.545,0.6775],[0.3825,0.6875],[0.265,0.6375],[0.2125,0.52],[0.215,0.41],[0.24,0.32],[0.28,0.2375],[0.33,0.17],[0.3625,0.1475],[0.3925,0.1425],[0.4475,0.1375],[0.485,0.1275],[0.5025,0.1175],[0.51,0.1175]],[[0.5525,0.1475],[0.6875,0.1825],[0.7425,0.245],[0.7825,0.3475],[0.765,0.475],[0.6625,0.59],[0.545,0.655],[0.385,0.6525],[0.3125,0.6025],[0.28,0.4675],[0.2925,0.35],[0.33,0.2775],[0.3525,0.2425],[0.3925,0.2025],[0.4325,0.1825],[0.485,0.17],[0.5225,0.165],[0.5525,0.1575],[0.56,0.1575],[0.56,0.1575]],[[0.5,0.1675],[0.605,0.17],[0.6825,0.2025],[0.7325,0.2675],[0.7625,0.3875],[0.7525,0.5175],[0.6975,0.6025],[0.6075,0.6475],[0.4625,0.66],[0.3475,0.61],[0.295,0.505],[0.2825,0.375],[0.31,0.2675],[0.36,0.1875],[0.41,0.1525],[0.4725,0.1475],[0.5175,0.1575],[0.5475,0.1625],[0.5525,0.1625],[0.5525,0.165]],[[0.4375,0.1525],[0.4775,0.1525],[0.5875,0.1825],[0.6425,0.255],[0.69,0.385],[0.6375,0.5825],[0.5425,0.68],[0.4375,0.7175],[0.3275,0.6875],[0.2625,0.5975],[0.2325,0.4875],[0.24,0.3175],[0.2575,0.25],[0.3125,0.2],[0.375,0.175],[0.4125,0.165],[0.4375,0.1575],[0.4575,0.155],[0.46,0.155],[0.4625,0.1525]],[[0.41,0.1475],[0.5425,0.145],[0.6275,0.1675],[0.7125,0.24],[0.7625,0.3625],[0.75,0.5125],[0.6625,0.6125],[0.4525,0.655],[0.38,0.6375],[0.3075,0.54],[0.2825,0.4175],[0.2975,0.3225],[0.345,0.2475],[0.375,0.2],[0.405,0.17],[0.4325,0.1525],[0.4625,0.145],[0.495,0.145],[0.5225,0.1475],[0.5325,0.15]],[[0.445,0.1375],[0.545,0.135],[0.6225,0.1575],[0.695,0.1975],[0.7325,0.2875],[0.7675,0.4275],[0.75,0.5725],[0.6075,0.705],[0.49,0.7075],[0.2875,0.65],[0.1725,0.5225],[0.1625,0.415],[0.21,0.2925],[0.2625,0.2275],[0.3225,0.1775],[0.37,0.1625],[0.4025,0.1575],[0.445,0.1575],[0.4925,0.155],[0.53,0.1525]],[[0.5825,0.1325],[0.7375,0.1675],[0.8,0.255],[0.81,0.44],[0.75,0.5975],[0.6425,0.685],[0.4825,0.7175],[0.3375,0.6775],[0.2425,0.595],[0.1875,0.495],[0.1775,0.3925],[0.2175,0.3125],[0.265,0.2475],[0.3225,0.1975],[0.365,0.1725],[0.42,0.1425],[0.4875,0.1],[0.59,0.0825],[0.6325,0.0825],[0.635,0.0825]],[[0.445,0.1675],[0.5625,0.16],[0.65,0.195],[0.72,0.295],[0.745,0.425],[0.7175,0.5625],[0.6125,0.6675],[0.475,0.7],[0.3525,0.685],[0.1625,0.61],[0.1125,0.5575],[0.105,0.475],[0.1625,0.3775],[0.2125,0.3175],[0.2925,0.2475],[0.3575,0.2025],[0.4275,0.1775],[0.475,0.1625],[0.4925,0.1575],[0.5025,0.1575]],[[0.515,0.155],[0.5725,0.1625],[0.71,0.2375],[0.78,0.3375],[0.795,0.4375],[0.77,0.5875],[0.665,0.6975],[0.5,0.7225],[0.4075,0.7025],[0.3025,0.6025],[0.2425,0.485],[0.2425,0.385],[0.285,0.2975],[0.325,0.2475],[0.38,0.2075],[0.4275,0.1875],[0.445,0.18],[0.5,0.1675],[0.55,0.165],[0.5725,0.1625]],[[0.4675,0.15],[0.5525,0.14],[0.6675,0.185],[0.7525,0.2675],[0.7925,0.37],[0.7925,0.4875],[0.6925,0.6175],[0.52,0.69],[0.36,0.6725],[0.2275,0.595],[0.1825,0.4775],[0.18,0.37],[0.2125,0.2975],[0.2725,0.2275],[0.355,0.175],[0.41,0.15],[0.4525,0.1425],[0.495,0.1375],[0.5125,0.1375],[0.515,0.1375]],[[0.5075,0.1375],[0.6,0.1375],[0.6825,0.1775],[0.7525,0.2575],[0.7875,0.3725],[0.7875,0.52],[0.6925,0.6375],[0.545,0.7],[0.395,0.66],[0.3025,0.5375],[0.2825,0.425],[0.31,0.2975],[0.3575,0.2125],[0.4025,0.1775],[0.46,0.1625],[0.52,0.1575],[0.555,0.1575],[0.5725,0.1575],[0.5725,0.1575],[0.5725,0.1575]]];
-
-
-let boxes = [[[0.4425,0.75],[0.46,0.7475],[0.545,0.7475],[0.6525,0.7475],[0.6825,0.7475],[0.68,0.78],[0.675,0.8875],[0.675,0.92],[0.665,0.93],[0.4825,0.9175],[0.41,0.925],[0.405,0.925],[0.4125,0.8375],[0.4275,0.7875],[0.4375,0.7575],[0.4375,0.7575],[0.4375,0.7575],[0.4375,0.7575],[0.4375,0.7575],[0.4375,0.7525]],[[0.4475,0.7375],[0.5425,0.74],[0.6525,0.74],[0.6925,0.74],[0.6925,0.7775],[0.69,0.86],[0.6925,0.915],[0.6875,0.915],[0.5875,0.9125],[0.4775,0.9125],[0.4525,0.9125],[0.4525,0.8875],[0.45,0.8075],[0.4525,0.7675],[0.4525,0.7525],[0.4525,0.7475],[0.4525,0.7475],[0.4525,0.7475],[0.4525,0.7475],[0.4525,0.7475]],[[0.4575,0.7425],[0.485,0.74],[0.6025,0.7425],[0.7025,0.7425],[0.7325,0.745],[0.7375,0.8175],[0.7425,0.9175],[0.7425,0.9225],[0.7225,0.92],[0.56,0.91],[0.4375,0.9125],[0.43,0.915],[0.4275,0.8825],[0.4375,0.8175],[0.445,0.77],[0.4475,0.7625],[0.45,0.7575],[0.45,0.7525],[0.45,0.75],[0.45,0.75]],[[0.43,0.7375],[0.4425,0.73],[0.565,0.73],[0.6575,0.73],[0.705,0.73],[0.7075,0.7325],[0.7075,0.82],[0.7025,0.8825],[0.7,0.8875],[0.66,0.885],[0.5475,0.8825],[0.425,0.8825],[0.4,0.8825],[0.3975,0.86],[0.395,0.81],[0.4075,0.77],[0.42,0.7425],[0.4225,0.7375],[0.4225,0.7375],[0.4225,0.7375]],[[0.3825,0.7025],[0.3875,0.7025],[0.56,0.7],[0.68,0.6975],[0.73,0.6925],[0.7325,0.6925],[0.7175,0.7575],[0.7025,0.8475],[0.7,0.8925],[0.6875,0.895],[0.58,0.875],[0.4675,0.87],[0.41,0.87],[0.4025,0.8725],[0.39,0.8075],[0.3875,0.76],[0.3825,0.7225],[0.38,0.7025],[0.38,0.6975],[0.38,0.6975]],[[0.4275,0.6975],[0.6425,0.6975],[0.75,0.6975],[0.7675,0.6975],[0.7675,0.72],[0.7625,0.8025],[0.76,0.865],[0.7525,0.8775],[0.665,0.86],[0.55,0.8575],[0.4725,0.8575],[0.435,0.86],[0.43,0.8625],[0.4175,0.7925],[0.415,0.7275],[0.415,0.71],[0.415,0.6975],[0.415,0.6975],[0.415,0.6975],[0.415,0.6975]],[[0.415,0.6975],[0.485,0.6925],[0.5925,0.6925],[0.7325,0.695],[0.7425,0.6975],[0.7425,0.73],[0.735,0.81],[0.7325,0.87],[0.73,0.895],[0.6925,0.8975],[0.6025,0.8925],[0.515,0.8925],[0.455,0.8875],[0.44,0.8875],[0.435,0.855],[0.4375,0.79],[0.43,0.7475],[0.42,0.705],[0.4175,0.7],[0.4175,0.6975]],[[0.4375,0.69],[0.5725,0.6875],[0.6725,0.6875],[0.7525,0.6875],[0.7625,0.6875],[0.7625,0.7075],[0.755,0.775],[0.7525,0.815],[0.7425,0.8475],[0.7275,0.8675],[0.66,0.86],[0.5625,0.855],[0.4725,0.85],[0.445,0.8525],[0.4375,0.8375],[0.4275,0.7675],[0.4225,0.7275],[0.425,0.7025],[0.4275,0.6975],[0.4275,0.6975]],[[0.4175,0.7125],[0.44,0.7075],[0.5625,0.7075],[0.6875,0.7075],[0.7575,0.7075],[0.76,0.7075],[0.7625,0.745],[0.7625,0.8125],[0.7575,0.8625],[0.7425,0.8875],[0.7325,0.8875],[0.645,0.8675],[0.5525,0.865],[0.4725,0.8675],[0.415,0.875],[0.4125,0.8675],[0.4125,0.8075],[0.4125,0.7675],[0.4125,0.74],[0.4125,0.7275]],[[0.41,0.7225],[0.42,0.72],[0.5575,0.72],[0.665,0.7225],[0.695,0.7225],[0.695,0.7275],[0.695,0.785],[0.7025,0.8575],[0.7025,0.9],[0.7,0.9025],[0.65,0.89],[0.54,0.8775],[0.4525,0.88],[0.435,0.885],[0.4175,0.86],[0.405,0.795],[0.4,0.7525],[0.4,0.7375],[0.4025,0.7225],[0.4025,0.7225]],[[0.405,0.7175],[0.47,0.7125],[0.6175,0.7125],[0.6875,0.7125],[0.715,0.7125],[0.715,0.7125],[0.705,0.765],[0.7025,0.815],[0.6975,0.865],[0.685,0.89],[0.6575,0.8875],[0.5575,0.8825],[0.4825,0.8825],[0.42,0.885],[0.415,0.8825],[0.415,0.815],[0.4075,0.7525],[0.405,0.7275],[0.405,0.7175],[0.405,0.7175]],[[0.405,0.7175],[0.4375,0.705],[0.5525,0.705],[0.66,0.705],[0.6875,0.705],[0.7,0.7075],[0.7025,0.725],[0.7,0.7925],[0.7,0.85],[0.695,0.8825],[0.6925,0.8875],[0.6625,0.8775],[0.5775,0.8675],[0.4825,0.8675],[0.3975,0.875],[0.3775,0.875],[0.3775,0.8375],[0.38,0.785],[0.3825,0.7475],[0.39,0.7275]],[[0.355,0.735],[0.3625,0.7325],[0.44,0.7275],[0.525,0.7275],[0.665,0.73],[0.6775,0.7325],[0.6775,0.7325],[0.675,0.7625],[0.6725,0.83],[0.6725,0.875],[0.6675,0.895],[0.6625,0.905],[0.6425,0.9025],[0.545,0.8875],[0.4625,0.8875],[0.4225,0.89],[0.4025,0.895],[0.375,0.8275],[0.365,0.7725],[0.3625,0.7575]],[[0.38,0.7475],[0.4125,0.7475],[0.47,0.7475],[0.545,0.745],[0.6425,0.745],[0.675,0.745],[0.6775,0.75],[0.685,0.8175],[0.695,0.8675],[0.695,0.9],[0.695,0.905],[0.635,0.8975],[0.54,0.895],[0.42,0.9075],[0.385,0.9075],[0.3725,0.8975],[0.3725,0.8525],[0.3775,0.815],[0.38,0.7925],[0.38,0.7675]],[[0.375,0.7625],[0.395,0.7575],[0.4825,0.7525],[0.56,0.7525],[0.6425,0.7525],[0.675,0.755],[0.685,0.7575],[0.6925,0.7875],[0.6925,0.875],[0.695,0.9025],[0.6825,0.9075],[0.6,0.9075],[0.5125,0.9075],[0.41,0.9075],[0.3775,0.9075],[0.3725,0.885],[0.38,0.84],[0.385,0.7875],[0.385,0.7575],[0.385,0.7575]],[[0.39,0.7525],[0.46,0.745],[0.5525,0.74],[0.6525,0.7375],[0.6625,0.7375],[0.6625,0.755],[0.65,0.815],[0.6475,0.8675],[0.645,0.8875],[0.64,0.9075],[0.6325,0.9075],[0.5925,0.8925],[0.5325,0.8875],[0.42,0.8875],[0.3775,0.8875],[0.3725,0.87],[0.3825,0.8225],[0.3925,0.7875],[0.3925,0.7675],[0.3925,0.76]],[[0.41,0.74],[0.4575,0.7375],[0.5125,0.7375],[0.625,0.7375],[0.65,0.7375],[0.6775,0.7425],[0.6825,0.7475],[0.6825,0.76],[0.6775,0.8125],[0.6725,0.8475],[0.6675,0.87],[0.6625,0.875],[0.5825,0.8675],[0.4925,0.8675],[0.4275,0.87],[0.415,0.87],[0.4125,0.8675],[0.4125,0.815],[0.41,0.7875],[0.405,0.7575]],[[0.405,0.7175],[0.4675,0.715],[0.575,0.715],[0.6475,0.7175],[0.7025,0.7175],[0.705,0.725],[0.705,0.7675],[0.7,0.8475],[0.6875,0.8875],[0.6775,0.9],[0.625,0.8975],[0.5525,0.8875],[0.47,0.8875],[0.4075,0.88],[0.3925,0.855],[0.385,0.8075],[0.3825,0.7675],[0.3925,0.7225],[0.395,0.71],[0.395,0.71]],[[0.395,0.7225],[0.4125,0.7175],[0.4775,0.7175],[0.5675,0.7175],[0.635,0.7175],[0.65,0.7175],[0.655,0.725],[0.66,0.7725],[0.66,0.8125],[0.655,0.8425],[0.65,0.855],[0.6175,0.8475],[0.5325,0.84],[0.475,0.84],[0.415,0.84],[0.375,0.84],[0.37,0.8175],[0.37,0.7775],[0.3725,0.7475],[0.375,0.7275]],[[0.37,0.7375],[0.3925,0.7325],[0.525,0.7275],[0.5825,0.7275],[0.6325,0.7275],[0.655,0.7275],[0.655,0.74],[0.655,0.7875],[0.6475,0.825],[0.645,0.855],[0.645,0.885],[0.6375,0.8925],[0.5725,0.8825],[0.455,0.8775],[0.3925,0.8775],[0.355,0.875],[0.3525,0.865],[0.3625,0.8125],[0.375,0.7775],[0.3825,0.7475]],[[0.39,0.7075],[0.4725,0.705],[0.5575,0.705],[0.67,0.715],[0.6925,0.7175],[0.6975,0.72],[0.695,0.78],[0.6925,0.845],[0.6925,0.89],[0.6925,0.8975],[0.6225,0.8875],[0.525,0.88],[0.445,0.8775],[0.4125,0.8775],[0.3975,0.8575],[0.395,0.7875],[0.3925,0.7425],[0.39,0.7025],[0.39,0.6925],[0.39,0.6925]],[[0.3925,0.73],[0.3975,0.7325],[0.5075,0.73],[0.5875,0.73],[0.6525,0.73],[0.7,0.7275],[0.71,0.7275],[0.71,0.7775],[0.6975,0.85],[0.6875,0.8875],[0.685,0.89],[0.615,0.885],[0.515,0.88],[0.445,0.88],[0.4225,0.88],[0.4175,0.85],[0.4125,0.7925],[0.41,0.7675],[0.4025,0.7425],[0.4025,0.7325]],[[0.4075,0.73],[0.4825,0.7275],[0.585,0.7275],[0.6975,0.7275],[0.7125,0.73],[0.7225,0.7325],[0.725,0.7675],[0.7225,0.8275],[0.7225,0.865],[0.7225,0.875],[0.715,0.8775],[0.61,0.86],[0.5525,0.86],[0.48,0.86],[0.445,0.86],[0.4325,0.8275],[0.4275,0.7775],[0.4225,0.7575],[0.415,0.7375],[0.415,0.73]],[[0.4125,0.7275],[0.425,0.7275],[0.4625,0.7275],[0.515,0.725],[0.5625,0.7225],[0.61,0.7225],[0.6425,0.7225],[0.6525,0.7425],[0.655,0.7975],[0.655,0.845],[0.655,0.875],[0.6275,0.88],[0.555,0.88],[0.4825,0.88],[0.44,0.88],[0.4225,0.855],[0.415,0.7975],[0.4125,0.7525],[0.4125,0.7275],[0.4125,0.7275]],[[0.4225,0.7425],[0.4825,0.7375],[0.5825,0.7325],[0.6425,0.7325],[0.6675,0.7325],[0.6775,0.74],[0.6775,0.795],[0.6725,0.845],[0.6675,0.8775],[0.665,0.8875],[0.6425,0.89],[0.5575,0.88],[0.4725,0.8775],[0.4275,0.875],[0.4125,0.86],[0.405,0.805],[0.405,0.7775],[0.4125,0.7475],[0.4125,0.7375],[0.4125,0.7375]],[[0.4125,0.7575],[0.4225,0.7475],[0.495,0.7425],[0.5825,0.7425],[0.6625,0.74],[0.695,0.74],[0.7025,0.7475],[0.7025,0.7975],[0.7025,0.845],[0.6975,0.865],[0.6925,0.87],[0.65,0.8675],[0.555,0.86],[0.4825,0.86],[0.41,0.86],[0.395,0.8625],[0.39,0.82],[0.3975,0.775],[0.4025,0.7575],[0.405,0.7475]],[[0.3775,0.7375],[0.4025,0.73],[0.4725,0.7275],[0.5725,0.7275],[0.6275,0.7275],[0.6425,0.7275],[0.6575,0.73],[0.66,0.7575],[0.66,0.8075],[0.65,0.8775],[0.6475,0.895],[0.605,0.8825],[0.53,0.875],[0.4375,0.87],[0.405,0.87],[0.39,0.855],[0.3825,0.8075],[0.3825,0.7725],[0.3875,0.75],[0.39,0.7475]],[[0.3475,0.75],[0.4325,0.745],[0.52,0.745],[0.59,0.75],[0.605,0.7525],[0.6125,0.78],[0.6125,0.835],[0.61,0.8775],[0.61,0.89],[0.56,0.8925],[0.4625,0.8775],[0.385,0.8775],[0.37,0.8775],[0.3625,0.8775],[0.3525,0.8275],[0.3475,0.7675],[0.35,0.7475],[0.35,0.7475],[0.35,0.7475],[0.3575,0.7525]],[[0.4,0.7475],[0.455,0.745],[0.5575,0.745],[0.675,0.745],[0.6925,0.7475],[0.7025,0.76],[0.7025,0.7975],[0.7,0.875],[0.7,0.915],[0.6975,0.92],[0.62,0.8975],[0.51,0.885],[0.44,0.885],[0.4225,0.89],[0.41,0.8825],[0.4125,0.8175],[0.4125,0.7775],[0.41,0.7475],[0.4075,0.745],[0.4075,0.745]],[[0.3575,0.755],[0.44,0.75],[0.565,0.75],[0.635,0.7525],[0.67,0.7525],[0.675,0.76],[0.6725,0.835],[0.6675,0.8925],[0.6675,0.9075],[0.6675,0.91],[0.6125,0.8975],[0.525,0.89],[0.455,0.8875],[0.42,0.8875],[0.39,0.885],[0.3525,0.865],[0.3475,0.8575],[0.355,0.8175],[0.3625,0.7625],[0.365,0.7375]]];
-
-let labels = [[0, 1],[0, 1],[0, 1],[0, 1],[0, 1],[0, 1],[0, 1],[0, 1],[0, 1],[0, 1],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],
-[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],
-[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],
-[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0],[1, 0]]
-
-let test =[[[0.2625,0.2075],[0.275,0.205],[0.3925,0.2175],[0.665,0.2175],[0.8325,0.2175],[0.8375,0.2175],[0.8375,0.2275],[0.815,0.3475],[0.8125,0.585],[0.805,0.675],[0.8,0.6775],[0.5425,0.695],[0.3425,0.7425],[0.3325,0.7425],[0.325,0.725],[0.2925,0.5575],[0.2775,0.4],[0.285,0.2875],[0.29,0.23],[0.29,0.2125]]];
-
 
 function setup() {
-  ml5.setBackend("webgl");
-  createCanvas(w, h);
-  frameRate(10);
+  ml5.setBackend('webgl')
+  let options = {
+    inputs: ['x', 'y'],
+    outputs: ['label'],
+    task: 'classification',
+    debug: 'true',
+    learningRate: 0.5
+  };
+  model = ml5.timeSeries(options);
+  createCanvas(600, 400);
   background(220);
-  classifer = ml5.timeSeries();
+  UI();
+  frameRate(60);
+  
 }
 
 function draw() {
-  if (keyIsDown(67) && pressedOnce){
-      sequence.push([mouseX/w,mouseY/h]);
-      ellipse(mouseX,mouseY,10);
-      if (sequence.length == 20){
-        pressedOnce = false;
-        datasets.push(sequence);
-        sequence = [];
-        console.log("finished");
-        background(220);
-      }
+  let datapoints = map(count,0,rec_duration*num_seq, 0,num_seq)
+  
+  if (mouseIsPressed && pressedOnce){
     
-  } else if (keyIsDown(66) && pressedOnce){
-      sequence.push([mouseX/w,mouseY/h]);
-      ellipse(mouseX,mouseY,10);
-      if (sequence.length == 20){
-        pressedOnce = false;
-        datasets.push(sequence);
-        sequence = [];
-        console.log("finished");
-        background(220);
-      }
-  } else if (keyIsDown(65) && pressedOnce){
-    sequence.push([mouseX/w,mouseY/h]);
-    ellipse(mouseX,mouseY,10);
-    if (sequence.length == 20){
+    line(pmouseX, pmouseY, mouseX,mouseY);
+    let inputs = {x: mouseX,y: mouseY};
+    count++;
+    
+    if (datapoints % 1 == 0){
+        // sequence.push(inputs);
+        sequence.push(inputs);
+    } 
+    
+    if (sequence.length == num_seq){
+      
+      
       pressedOnce = false;
-      datasets.push(sequence);
+      count = 0
+      
+      if (state == 'collection'){
+        let target = {label: curr_shape};
+        background(220);
+        text("Recording: " + curr_shape, 50,50);
+        // console.log(sequence, target)
+        options = {inputLabels:['x','y']}
+        model.addData(sequence, target, options);
+      } else if (state == 'prediction'){
+        background(220);
+        model.classify(sequence, gotResults)
+      } else if (state == 'training') {
+        background(220);
+        text("You cannot record while training");
+      }
+
       sequence = [];
-      console.log("finished");
-      background(220);
-
-      tensortrial = classifer.toTensors(datasets,datasets);
-      predictions = classifer.predict(tensortrial[0]);
-      console.log(type[predictions], 'dito', predictions);
-
-      text(type[predictions],200,200);
-      // console.log(predictions[predictions.length-1])
     }
   } 
-  
-  if (datasets.length == 10){
-    file = JSON.stringify(datasets);
-    console.log(file);
-  }
 }
 
-function keyReleased(){
-  pressedOnce = true;
+function gotResults(results) {
+  // if (error) {
+  //   console.log(error);
+  // }
+  // console.log('hello', results);
+  stroke(0);
+  fill(0, 0, 255, 100);
+  let label = results[0].label;
+  text("Prediction: " + label, 50,50);
+  // let label = error[0].label;
+
 }
 
 function keyPressed(){
-    if (key == 't'){
-      classifer.createArchitecture();
-      console.log('done architecture');
-      classifer.compileModel();
-      console.log('done compiling the thing');
-      classifer.summarizeModel();
-      console.log('done summarizing');
-      alldata = circle.concat(boxes);
-      tensorData = classifer.toTensors(alldata,labels);
-      console.log('conversion done');
-      console.log(tensorData[0]);
-      console.log(tensorData[1]);
-
-      classifer.fitModel(tensorData[0],tensorData[1]);
-
-    } else if (key =='p'){
-        classifer.modelSummary();
-    }
+  if (key == 's') {
+    model.saveData('trial');
+  } else if (key == 'd'){
+    console.log(model.getData());
   }
+}
+
+function mouseReleased(){
+  pressedOnce = true;
+}
+
+function UI(){
+  
+  textSize(20)
+  
+  rec_circle = createButton('Record Circle');
+  rec_circle.mouseClicked(recordCircle);
+  rec_circle.style("font-family", "Georgia");
+  rec_circle.style("font-size", "20px");
+  
+  rec_square = createButton('Record Square');
+  rec_square.mouseClicked(recordSquare);
+  rec_square.style("font-family", "Georgia");
+  rec_square.style("font-size", "20px");
+  
+  train_but = createButton('Train Model');
+  train_but.mouseClicked(trainModel);
+  train_but.style("font-family", "Georgia");
+  train_but.style("font-size", "20px");
+  
+  pred_sha = createButton('Predict Shape');
+  pred_sha.mouseClicked(predictShape);
+  pred_sha.style("font-family", "Georgia");
+  pred_sha.style("font-size", "20px");
+  
+  function recordCircle(){
+    background(220);
+    state = 'collection'
+    curr_shape = 'Circle'
+    text("Recording: Circle", 50,50);
+    rec_circle.style("background-color",'#f0f0f0')
+    rec_square.style('background-color', '');
+    pred_sha.style('background-color', '');
+  }
+
+  function recordSquare(){
+    background(220);
+    state = 'collection'
+    curr_shape = 'Square'
+    text("Recording: Square", 50,50);
+    rec_square.style("background-color",'#f0f0f0')
+    rec_circle.style('background-color', '');
+    pred_sha.style('background-color', '');
+  }
+
+  function trainModel(){
+    model.createArchitecture();
+    model.compileModel();
+    model.summarizeModel();
+    background(220);
+    state = 'training';
+    text("Training...", 50,50);
+    model.normalizeData();
+    let options = {
+      epochs: 100
+    }
+    model.train(options,whileTraining,finishedTraining);
+  }
+  
+  function whileTraining(epoch, loss) {
+    console.log(epoch);
+  }
+  
+  function finishedTraining() {
+    console.log('finished training.');
+    state = 'prediction';
+  }
+
+  function predictShape(){
+    background(220);
+    state = 'prediction'
+    text("Predicting Shape...", 50,50);
+    pred_sha.style("background-color",'#f0f0f0')
+    rec_square.style('background-color', '');
+    rec_circle.style('background-color', '');
+
+
+  }
+}
+
