+++ src/LSTM/timeSeries.js
@@ -2,6 +2,17 @@ import * as tf from "@tensorflow/tfjs";
 import { saveBlob } from "../utils/io";
 import { randomGaussian } from "../utils/random";
 
+
+
+/*
+
+Things changed from neural network class:
+
+1. No neuro evolution
+
+
+*/
+
 class NeuralNetwork {
   constructor() {
     // flags
@@ -235,70 +246,6 @@ class NeuralNetwork {
    */
   dispose() {
     this.model.dispose();
-  }
-
-  // NeuroEvolution Functions
-
-  /**
-   * mutate the weights of a model
-   * @param {*} rate
-   * @param {*} mutateFunction
-   */
-
-  mutate(rate = 0.1, mutateFunction) {
-    tf.tidy(() => {
-      const weights = this.model.getWeights();
-      const mutatedWeights = [];
-      for (let i = 0; i < weights.length; i += 1) {
-        const tensor = weights[i];
-        const { shape } = weights[i];
-        // TODO: Evaluate if this should be sync or not
-        const values = tensor.dataSync().slice();
-        for (let j = 0; j < values.length; j += 1) {
-          if (Math.random() < rate) {
-            if (mutateFunction) {
-              values[j] = mutateFunction(values[j]);
-            } else {
-              values[j] = Math.min(
-                Math.max(values[j] + randomGaussian(), -1),
-                1
-              );
-            }
-          }
-        }
-        const newTensor = tf.tensor(values, shape);
-        mutatedWeights[i] = newTensor;
-      }
-      this.model.setWeights(mutatedWeights);
-    });
-  }
-
-  /**
-   * create a new neural network with crossover
-   * @param {*} other
-   */
-  crossover(other) {
-    return tf.tidy(() => {
-      const weightsA = this.model.getWeights();
-      const weightsB = other.model.getWeights();
-      const childWeights = [];
-      for (let i = 0; i < weightsA.length; i += 1) {
-        const tensorA = weightsA[i];
-        const tensorB = weightsB[i];
-        const { shape } = weightsA[i];
-        // TODO: Evaluate if this should be sync or not
-        const valuesA = tensorA.dataSync().slice();
-        const valuesB = tensorB.dataSync().slice();
-        for (let j = 0; j < valuesA.length; j += 1) {
-          if (Math.random() < 0.5) {
-            valuesA[j] = valuesB[j];
-          }
-        }
-        const newTensor = tf.tensor(valuesA, shape);
-        childWeights[i] = newTensor;
-      }
-      this.model.setWeights(childWeights);
-    });
   }
 }
 export default NeuralNetwork;
